# Conflicts:
#	src/main/java/com/zf/controller/admin/ManageUserController.java
#	src/main/java/com/zf/service/SysUserService.java
#	src/main/java/com/zf/service/impl/CaseContentServiceImpl.java
#	src/main/java/com/zf/service/impl/PersonalCardServiceImpl.java

+++ src/main/java/com/zf/mapper/PersonalCardMapper.java
@@ -22,5 +22,5 @@ public interface PersonalCardMapper extends BaseMapper<PersonalCardVo> {
    * @param id
    * @return
    */
-  PersonalCardVo selectPersonalCardById(@Param("id") Integer id);
+  PersonalCardVo selectPersonalCardById(@Param("id") Long id);
 }
